refactor files (mainly from utils folder)
