+++ src/components/police-components/police-radio/Display.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useState, useRef } from "react"
 import { Radiooption } from "./Radiooption"
 
-import { radioDisplayContent } from "../../../utils/policeRadioDisplayContent";
+import { policeRadioMenuContent } from "../../../data/policeRadioMenuContent.js";
 
 export function Display () {
     const [activeOptionIndex, setActiveOptionIndex] = useState(0);
@@ -9,9 +9,9 @@ export function Display () {
 
     function onKeyDown (e) {
         if (e.key === "ArrowUp") {
-            setActiveOptionIndex(prev => prev > 0 ? prev - 1 : radioDisplayContent.length - 1)
+            setActiveOptionIndex(prev => prev > 0 ? prev - 1 : policeRadioMenuContent.length - 1)
         } else if (e.key === "ArrowDown") {
-            setActiveOptionIndex(prev => prev < radioDisplayContent.length-1 ? prev + 1 : 0)
+            setActiveOptionIndex(prev => prev < policeRadioMenuContent.length-1 ? prev + 1 : 0)
         }
     };
 
@@ -24,7 +24,7 @@ return (
         className="bg-orange-100 tracking-tighter text-yellow-300 font-mono mt-48 ml-7 shadow-xl w-33"
         role="listbox"
     >
-        { radioDisplayContent.map((option,i) => (
+        { policeRadioMenuContent.map((option,i) => (
             <div
                 key={option.label}
                 className="mb-1 outline-none">
