+++ src/components/textboxes/BaseTextbox.jsx
@@ -1,5 +1,5 @@
 import {useEffect} from "react";
-import { closeTextbox } from "../../utils/closeTextbox";
+import { useCloseTextbox } from "../../hooks/useCloseTextbox.jsx";
 
 import { useGuiVisibilityStatesStore } from "../../store";
 
@@ -25,7 +25,7 @@ export const BaseTextbox = ({
     const textboxesAreVisible = useGuiVisibilityStatesStore(state => state.textboxesVisible);
 
     const handleClose = () => {
-        closeTextbox(setTextboxVisibililty,onClose);
+        useCloseTextbox(setTextboxVisibililty,onClose);
     };
 
     useEffect(() => {
