+++ src/data/policeRadioMenuContent.js
@@ -1,0 +1,18 @@
+export const policeRadioMenuContent = [
+    {
+        label: "Call tow truck",
+        initSpawn: true
+    },
+    {
+        label: "Call Patrol",
+        initSpawn: true
+    },
+    {
+        label: "Lookup driver",
+        initSpawn: false
+    },
+    {
+        label: "Lookup Car",
+        initSpawn: false
+    }
+]
