+++ src/hooks/useVehicleEntityGenerator.jsx
@@ -1,12 +1,8 @@
 import { useEffect } from "react";
-
-import { generateCarProfile } from "../utils/profileGenerators/carProfileGenerator.js";
-
 import { generateUUID, randInt } from "three/src/math/MathUtils.js";
 
 import { useCarStore, useNpcStore } from "../store.js";
-import { generateDriverProfile } from "../utils/profileGenerators/driverProfileGenerator.js";
-import { basicNpcProfile } from "../utils/generators/basicNpcProfileGenerator.js";
+import { basicNpcProfile } from "../utils/generators/entityProfileGenerators/basicNpcProfileGenerator.js";
 
 export function useVehicleEntityGenerator (direction, lane) {
     const addCar = useCarStore((state) => state.addCar);
