--- src/utils/SpawnHandler.js
@@ -1,9 +1,0 @@
-
-
-export class SpawnHandler {
-    constructor (spawnInfo) {
-        this.spawnInfo = spawnInfo;
-    }
-
-    
-}
