+++ src/utils/generators/entityProfileGenerators/fakeNpcProfileGenerator.js
